Revert changes

+++ test/libsolidity/semanticTests/abiEncoderV1/return_dynamic_types_cross_call_advanced.sol
@@ -3,8 +3,8 @@ contract C {
 		a = "1234567890123456789012345678901234567890";
 		b = uint(-1);
 		c = new bytes20[](4);
-		c[0] = bytes20(uint256(1234));
-		c[3] = bytes20(uint256(6789));
+		c[0] = bytes20(uint160(1234));
+		c[3] = bytes20(uint160(6789));
 		d = 0x1234;
 	}
 	function f() public returns (bytes memory, uint, bytes20[] memory, uint) {
